Advance-Vidder-Quiz-Bot/handlers/external_integration_handlers.py
# Powered by Viddertech

import logging
from telegram import Update
from telegram.ext import ContextTypes

logger = logging.getLogger(__name__)

async def login_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """Handles the /login command for TestBook integration."""
    await update.message.reply_text(
        "**TestBook Integration**\n\n"
        "This feature is designed to connect with your TestBook account to import quizzes directly from test links.\n\n"
        "**Status:** Under Development\n"
        "Final integration is pending access to the official TestBook API. Stay tuned for updates!",
        parse_mode='Markdown'
    )

async def lang_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """Handles the /lang command for TestBook integration."""
    await update.message.reply_text(
        "**Language Selection**\n\n"
        "This feature will allow you to select your preferred language for quizzes imported from TestBook.\n\n"
        "**Status:** Under Development\n"
        "This will be activated alongside the main TestBook integration.",
        parse_mode='Markdown'
    )