Advance-Vidder-Quiz-Bot/requirements.txt
# --- Core Libraries ---
python-telegram-bot[ext]
sqlalchemy
python-dotenv

# --- Integrations ---
telethon
openai
pytesseract
pdf2image
requests
beautifulsoup4

# --- Database Drivers (optional but good practice) ---
# For PostgreSQL
# psycopg2-binary

# For MySQL
# pymysql